refactor(api-rs): move session streaming into runtime
