+++ services/api-rs/crates/centaur-api-server/Cargo.toml
@@ -7,11 +7,11 @@ repository.workspace = true
 
 [dependencies]
 axum.workspace = true
-bytes.workspace = true
 centaur-sandbox-agent-k8s.workspace = true
 centaur-sandbox-core.workspace = true
 centaur-sandbox-local.workspace = true
 centaur-session-core.workspace = true
+centaur-session-runtime.workspace = true
 centaur-session-sqlx.workspace = true
 clap.workspace = true
 futures-util.workspace = true
@@ -20,7 +20,6 @@ serde.workspace = true
 serde_json.workspace = true
 sqlx.workspace = true
 thiserror.workspace = true
-tokio = { workspace = true, features = ["io-util", "macros", "net", "rt-multi-thread", "signal", "time"] }
-tokio-util = { workspace = true, features = ["codec"] }
+tokio = { workspace = true, features = ["macros", "net", "rt-multi-thread", "signal"] }
 tracing.workspace = true
 tracing-subscriber = { workspace = true, features = ["env-filter", "fmt", "json"] }
