+++ services/api-rs/crates/centaur-session-runtime/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "centaur-session-runtime"
+version = "0.1.0"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+centaur-sandbox-core.workspace = true
+centaur-session-core.workspace = true
+centaur-session-sqlx.workspace = true
+futures-util.workspace = true
+serde_json.workspace = true
+thiserror.workspace = true
+tokio = { workspace = true, features = ["io-util", "sync", "time"] }
+tokio-util = { workspace = true, features = ["codec"] }
+tracing.workspace = true
